refactor: move timezone handling logic to a separate module
